requirements.txt
pytorch==1.0
PyYAML>=5.1
pandas
scikit-learn
numpy
h5py
tqdm
